+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 6.2.0
+* enable http/s sources for templates
+* enable http/s sources for models
+
 # 6.1.0
 * add additional functions to modelFuncs
 * fix random date generation in a way, that it produces ISO strings